# Refactor use of FuncForPC to leverage new CallersFrames API.