Sorry to tack this on to another pending PR, I can create a new branch with this modification cherry-picked but I feel that the other changes were necessary regardless and this kind of eliminates the need for them entirely. 

I believe this is what you were looking for in #10. The performance for this took a huge hit because `runtime.CallersFrames` allocates way more than simply grabbing the function pointer for the PC for obvious reasons. It does simplify much of the internal PC trickery, so perhaps the trade off is worth it.

There's also the possibility that my specific implementation may be less-than optimal as I'm not claiming to be a paragon of performant code here.

```
name                   old time/op  new time/op   delta
CallVFmt-8              498ns ± 1%    507ns ± 2%      ~     (p=0.057 n=4+4)
CallPlusVFmt-8          650ns ± 2%    658ns ± 4%      ~     (p=0.743 n=4+4)
CallSharpVFmt-8         504ns ± 5%    518ns ± 5%      ~     (p=0.486 n=4+4)
CallSFmt-8              451ns ± 1%    475ns ± 5%    +5.32%  (p=0.029 n=4+4)
CallPlusSFmt-8          587ns ± 1%    613ns ± 2%    +4.39%  (p=0.029 n=4+4)
CallSharpSFmt-8         438ns ± 1%    455ns ± 2%    +3.94%  (p=0.029 n=4+4)
CallDFmt-8              404ns ± 1%    403ns ± 1%      ~     (p=0.314 n=4+4)
CallNFmt-8              379ns ± 1%    389ns ± 4%      ~     (p=0.200 n=4+4)
CallPlusNFmt-8          347ns ± 1%    353ns ± 1%      ~     (p=0.086 n=4+4)
Caller-8                372ns ± 0%    616ns ± 3%   +65.68%  (p=0.029 n=4+4)
Trace-8                 850ns ± 4%   2306ns ± 1%  +171.40%  (p=0.029 n=4+4)
Trace10-8              2.82µs ± 1%   7.44µs ± 7%  +164.17%  (p=0.029 n=4+4)
Trace50-8              4.52µs ± 1%  16.28µs ± 1%  +260.27%  (p=0.029 n=4+4)
Trace100-8             7.83µs ± 1%  30.05µs ± 6%  +283.89%  (p=0.029 n=4+4)
CallerAndVFmt-8         918ns ± 2%   1206ns ± 7%   +31.25%  (p=0.029 n=4+4)
TraceAndVFmt-8         2.65µs ± 4%   4.19µs ± 4%   +58.23%  (p=0.029 n=4+4)
Trace10AndVFmt-8       9.38µs ± 1%  13.50µs ± 3%   +43.89%  (p=0.029 n=4+4)
RuntimeCaller-8         391ns ± 1%    404ns ± 6%      ~     (p=0.314 n=4+4)
RuntimeCallerAndFmt-8   714ns ± 2%    751ns ± 7%      ~     (p=0.114 n=4+4)
FuncForPC-8            9.27ns ± 1%  10.47ns ± 6%   +12.97%  (p=0.029 n=4+4)
FuncFileLine-8          123ns ± 1%    129ns ± 7%      ~     (p=0.086 n=4+4)
```