@dkushner 2017-06-13
Here is the new benchstat output after making the performance improvements we spoke about:

```
name                   old time/op  new time/op  delta
CallVFmt-8              560ns ± 2%   682ns ± 7%  +21.94%  (p=0.029 n=4+4)
CallPlusVFmt-8          712ns ± 2%   754ns ± 3%   +5.86%  (p=0.029 n=4+4)
CallSharpVFmt-8         553ns ± 5%   643ns ± 4%  +16.33%  (p=0.029 n=4+4)
CallSFmt-8              495ns ± 3%   586ns ± 2%  +18.33%  (p=0.029 n=4+4)
CallPlusSFmt-8          650ns ± 2%   689ns ± 1%   +6.00%  (p=0.029 n=4+4)
CallSharpSFmt-8         478ns ± 1%   586ns ± 1%  +22.52%  (p=0.029 n=4+4)
CallDFmt-8              428ns ± 2%   534ns ± 3%  +24.77%  (p=0.029 n=4+4)
CallNFmt-8              404ns ± 2%   605ns ± 2%  +49.66%  (p=0.029 n=4+4)
CallPlusNFmt-8          375ns ± 2%   579ns ± 6%  +54.33%  (p=0.029 n=4+4)
Caller-8                660ns ± 3%   408ns ± 8%  -38.28%  (p=0.029 n=4+4)
Trace-8                2.65µs ±24%  0.81µs ± 3%  -69.38%  (p=0.029 n=4+4)
Trace10-8              7.71µs ± 7%  2.52µs ± 2%  -67.34%  (p=0.029 n=4+4)
Trace50-8              17.5µs ± 2%   4.0µs ± 1%  -77.11%  (p=0.029 n=4+4)
Trace100-8             32.2µs ± 2%   6.9µs ± 2%  -78.47%  (p=0.029 n=4+4)
CallerAndVFmt-8        1.29µs ± 2%  1.08µs ± 3%  -16.51%  (p=0.029 n=4+4)
TraceAndVFmt-8         4.60µs ±10%  3.13µs ± 3%  -32.00%  (p=0.029 n=4+4)
Trace10AndVFmt-8       14.5µs ± 5%  11.1µs ±10%  -23.36%  (p=0.029 n=4+4)
RuntimeCaller-8         432ns ± 2%   434ns ± 5%     ~     (p=0.686 n=4+4)
RuntimeCallerAndFmt-8   782ns ± 3%   800ns ± 5%     ~     (p=0.629 n=4+4)
FuncForPC-8            11.2ns ± 4%  11.0ns ± 3%     ~     (p=0.686 n=4+4)
FuncFileLine-8          137ns ± 2%   143ns ±15%     ~     (p=0.657 n=4+4)
```

Note that this is against the previous, non-optimized version and not the original version that used `FuncForPC`.