@ChrisHines 2017-06-14
This looks good in general, but I'm not ready to merge it yet. It may still need some tweaks, but it will require some time to think through the best next step.

I ran a benchmark comparison of this PR against v1.5.3 and got the numbers shown below. It looks like the overall work is roughly the same it has just been shifted from the initial capture towards the formatting methods, which is actually a good thing. So performance is fine now.

Digging through the new code, though, I'm a little nervous about how it is using `CallersFrames`. I don't think the new code handles the `sigpanic` special case properly because it doesn't always use `runtime.Frames.Next` to traverse the entry prior to the one of interest. Take a look at the code for `runtime.Frames` and `runtime.Frames.Next` for the details about how that works.

I also just reviewed https://github.com/golang/go/issues/19348, which is the discussion about how Go will implement mid-stack inlining while preserving stack traces. The changes to  `runtime.Callers` and `runtime.CallersFrames` will change somewhat at that point and this package needs to work with those changes, but I don't understand all the implications yet.

### Benchmarks vs. v1.5.3
```
name                   old time/op    new time/op    delta
CallVFmt-8                429ns ± 2%     518ns ± 5%  +20.54%          (p=0.008 n=5+5)
CallPlusVFmt-8            551ns ± 1%     632ns ± 3%  +14.75%          (p=0.008 n=5+5)
CallSharpVFmt-8           410ns ± 0%     539ns ± 6%  +31.64%          (p=0.016 n=4+5)
CallSFmt-8                382ns ± 2%     493ns ± 6%  +28.87%          (p=0.008 n=5+5)
CallPlusSFmt-8            499ns ± 2%     595ns ± 2%  +19.33%          (p=0.008 n=5+5)
CallSharpSFmt-8           367ns ± 1%     493ns ± 2%  +34.24%          (p=0.008 n=5+5)
CallDFmt-8                336ns ± 1%     444ns ± 5%  +32.24%          (p=0.008 n=5+5)
CallNFmt-8                321ns ± 1%     519ns ± 3%  +61.88%          (p=0.008 n=5+5)
CallPlusNFmt-8            295ns ± 1%     499ns ± 1%  +69.29%          (p=0.008 n=5+5)
Caller-8                  307ns ± 4%     332ns ± 8%   +8.07%          (p=0.016 n=5+5)
Trace-8                   701ns ± 3%     700ns ± 4%     ~             (p=0.690 n=5+5)
Trace10-8                2.44µs ± 9%    2.08µs ± 2%  -14.66%          (p=0.008 n=5+5)
Trace50-8                4.18µs ± 4%    3.17µs ± 2%  -23.98%          (p=0.008 n=5+5)
Trace100-8               6.97µs ± 3%    5.58µs ± 1%  -19.94%          (p=0.008 n=5+5)
CallerAndVFmt-8           886ns ± 3%     878ns ± 1%     ~             (p=0.841 n=5+5)
TraceAndVFmt-8           2.30µs ± 6%    2.53µs ± 2%   +9.73%          (p=0.008 n=5+5)
Trace10AndVFmt-8         8.12µs ± 2%    9.34µs ± 3%  +15.10%          (p=0.008 n=5+5)
RuntimeCaller-8           353ns ± 1%     354ns ± 7%     ~             (p=0.413 n=5+5)
RuntimeCallerAndFmt-8     656ns ± 1%     682ns ± 7%     ~             (p=0.151 n=5+5)
FuncForPC-8              9.07ns ± 0%    8.45ns ± 6%   -6.77%          (p=0.016 n=4+5)
FuncFileLine-8            115ns ± 1%     111ns ± 2%   -3.31%          (p=0.008 n=5+5)

name                   old alloc/op   new alloc/op   delta
CallVFmt-8                40.0B ± 0%     40.0B ± 0%     ~     (all samples are equal)
CallPlusVFmt-8            72.0B ± 0%     72.0B ± 0%     ~     (all samples are equal)
CallSharpVFmt-8           88.0B ± 0%     88.0B ± 0%     ~     (all samples are equal)
CallSFmt-8                32.0B ± 0%     32.0B ± 0%     ~     (all samples are equal)
CallPlusSFmt-8            64.0B ± 0%     64.0B ± 0%     ~     (all samples are equal)
CallSharpSFmt-8           80.0B ± 0%     80.0B ± 0%     ~     (all samples are equal)
CallDFmt-8                24.0B ± 0%     24.0B ± 0%     ~     (all samples are equal)
CallNFmt-8                48.0B ± 0%     48.0B ± 0%     ~     (all samples are equal)
CallPlusNFmt-8            80.0B ± 0%     80.0B ± 0%     ~     (all samples are equal)
Caller-8                 0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
Trace-8                   64.0B ± 0%     32.0B ± 0%  -50.00%          (p=0.008 n=5+5)
Trace10-8                  240B ± 0%      128B ± 0%  -46.67%          (p=0.008 n=5+5)
Trace50-8                  896B ± 0%      448B ± 0%  -50.00%          (p=0.008 n=5+5)
Trace100-8               1.79kB ± 0%    0.90kB ± 0%  -50.00%          (p=0.008 n=5+5)
CallerAndVFmt-8           40.0B ± 0%     40.0B ± 0%     ~     (all samples are equal)
TraceAndVFmt-8             192B ± 0%      192B ± 0%     ~     (all samples are equal)
Trace10AndVFmt-8           640B ± 0%      640B ± 0%     ~     (all samples are equal)
RuntimeCaller-8          0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
RuntimeCallerAndFmt-8     40.0B ± 0%     40.0B ± 0%     ~     (all samples are equal)
FuncForPC-8              0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
FuncFileLine-8           0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)

name                   old allocs/op  new allocs/op  delta
CallVFmt-8                 3.00 ± 0%      4.00 ± 0%  +33.33%          (p=0.008 n=5+5)
CallPlusVFmt-8             3.00 ± 0%      4.00 ± 0%  +33.33%          (p=0.008 n=5+5)
CallSharpVFmt-8            3.00 ± 0%      4.00 ± 0%  +33.33%          (p=0.008 n=5+5)
CallSFmt-8                 2.00 ± 0%      3.00 ± 0%  +50.00%          (p=0.008 n=5+5)
CallPlusSFmt-8             2.00 ± 0%      3.00 ± 0%  +50.00%          (p=0.008 n=5+5)
CallSharpSFmt-8            2.00 ± 0%      3.00 ± 0%  +50.00%          (p=0.008 n=5+5)
CallDFmt-8                 2.00 ± 0%      3.00 ± 0%  +50.00%          (p=0.008 n=5+5)
CallNFmt-8                 2.00 ± 0%      3.00 ± 0%  +50.00%          (p=0.008 n=5+5)
CallPlusNFmt-8             2.00 ± 0%      3.00 ± 0%  +50.00%          (p=0.008 n=5+5)
Caller-8                  0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
Trace-8                    1.00 ± 0%      1.00 ± 0%     ~     (all samples are equal)
Trace10-8                  1.00 ± 0%      1.00 ± 0%     ~     (all samples are equal)
Trace50-8                  1.00 ± 0%      1.00 ± 0%     ~     (all samples are equal)
Trace100-8                 1.00 ± 0%      1.00 ± 0%     ~     (all samples are equal)
CallerAndVFmt-8            3.00 ± 0%      4.00 ± 0%  +33.33%          (p=0.008 n=5+5)
TraceAndVFmt-8             10.0 ± 0%      14.0 ± 0%  +40.00%          (p=0.008 n=5+5)
Trace10AndVFmt-8           32.0 ± 0%      47.0 ± 0%  +46.88%          (p=0.008 n=5+5)
RuntimeCaller-8           0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
RuntimeCallerAndFmt-8      3.00 ± 0%      3.00 ± 0%     ~     (all samples are equal)
FuncForPC-8               0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
FuncFileLine-8            0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
```