@dkushner 2017-07-31
Runtime and allocation benchmarks for the newest commits. I kind of decimated the previous changes by doing something very very stupid trying to resync my local. 

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCallVFmt-8                379           689           +81.79%
BenchmarkCallPlusVFmt-8            530           776           +46.42%
BenchmarkCallSharpVFmt-8           407           689           +69.29%
BenchmarkCallSFmt-8                350           630           +80.00%
BenchmarkCallPlusSFmt-8            472           712           +50.85%
BenchmarkCallSharpSFmt-8           344           622           +80.81%
BenchmarkCallDFmt-8                312           599           +91.99%
BenchmarkCallNFmt-8                287           647           +125.44%
BenchmarkCallPlusNFmt-8            264           622           +135.61%
BenchmarkCaller-8                  339           288           -15.04%
BenchmarkTrace-8                   741           626           -15.52%
BenchmarkTrace10-8                 2601          1982          -23.80%
BenchmarkTrace50-8                 5147          3343          -35.05%
BenchmarkTrace100-8                9201          5459          -40.67%
BenchmarkCallerAndVFmt-8           806           1048          +30.02%
BenchmarkTraceAndVFmt-8            2321          4005          +72.55%
BenchmarkTrace10AndVFmt-8          7824          12383         +58.27%
BenchmarkRuntimeCaller-8           325           327           +0.62%
BenchmarkRuntimeCallerAndFmt-8     625           621           -0.64%
BenchmarkFuncForPC-8               7.50          7.04          -6.13%
BenchmarkFuncFileLine-8            101           101           +0.00%

benchmark                          old allocs     new allocs     delta
BenchmarkCallVFmt-8                3              4              +33.33%
BenchmarkCallPlusVFmt-8            3              4              +33.33%
BenchmarkCallSharpVFmt-8           3              4              +33.33%
BenchmarkCallSFmt-8                2              3              +50.00%
BenchmarkCallPlusSFmt-8            2              3              +50.00%
BenchmarkCallSharpSFmt-8           2              3              +50.00%
BenchmarkCallDFmt-8                2              3              +50.00%
BenchmarkCallNFmt-8                2              3              +50.00%
BenchmarkCallPlusNFmt-8            2              3              +50.00%
BenchmarkCaller-8                  0              0              +0.00%
BenchmarkTrace-8                   1              1              +0.00%
BenchmarkTrace10-8                 1              1              +0.00%
BenchmarkTrace50-8                 1              1              +0.00%
BenchmarkTrace100-8                1              1              +0.00%
BenchmarkCallerAndVFmt-8           3              4              +33.33%
BenchmarkTraceAndVFmt-8            10             15             +50.00%
BenchmarkTrace10AndVFmt-8          32             38             +18.75%
BenchmarkRuntimeCaller-8           0              0              +0.00%
BenchmarkRuntimeCallerAndFmt-8     3              3              +0.00%
BenchmarkFuncForPC-8               0              0              +0.00%
BenchmarkFuncFileLine-8            0              0              +0.00%

benchmark                          old bytes     new bytes     delta
BenchmarkCallVFmt-8                40            56            +40.00%
BenchmarkCallPlusVFmt-8            72            88            +22.22%
BenchmarkCallSharpVFmt-8           104           120           +15.38%
BenchmarkCallSFmt-8                32            48            +50.00%
BenchmarkCallPlusSFmt-8            64            80            +25.00%
BenchmarkCallSharpSFmt-8           96            112           +16.67%
BenchmarkCallDFmt-8                24            40            +66.67%
BenchmarkCallNFmt-8                48            64            +33.33%
BenchmarkCallPlusNFmt-8            80            96            +20.00%
BenchmarkCaller-8                  0             0             +0.00%
BenchmarkTrace-8                   64            80            +25.00%
BenchmarkTrace10-8                 240           256           +6.67%
BenchmarkTrace50-8                 896           896           +0.00%
BenchmarkTrace100-8                1792          1792          +0.00%
BenchmarkCallerAndVFmt-8           40            56            +40.00%
BenchmarkTraceAndVFmt-8            192           456           +137.50%
BenchmarkTrace10AndVFmt-8          640           1152          +80.00%
BenchmarkRuntimeCaller-8           0             0             +0.00%
BenchmarkRuntimeCallerAndFmt-8     40            40            +0.00%
BenchmarkFuncForPC-8               0             0             +0.00%
BenchmarkFuncFileLine-8            0             0             +0.00%
```